chg: Move getChainName() from util/chain/ to util/
